benchmark-tests/requirements-test.txt
nltk
unbabel-comet
requests